% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzbuzz.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{FizzBuzz function}
\usage{
fizzbuzz(input)
}
\arguments{
\item{input}{a numeric vector}
}
\value{
fizzbuzz returns a character vector
}
\description{
a
}
\examples{
fizzbuzz(c(1,3,5,15))

}
