% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzbuzz.R
\name{divisibleBy}
\alias{divisibleBy}
\title{divisibleBy}
\usage{
divisibleBy(range, triggers)
}
\arguments{
\item{triggers}{}
}
\value{

}
\description{
A program that prints the numbers from 1 to 100. But for multiples of
three prints "Fizz" instead of the number and for the multiples of five print
"Buzz". For numbers which are multiples of both three and five print "FizzBuzz".
}
