% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzbuzz.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{The fizzbuzz function}
\usage{
fizzbuzz(number)
}
\arguments{
\item{number}{The input number}
}
\value{
string or number
}
\description{
Returns 'Fizz' if the input is a multiple of 3
Returns 'Buzz' if the input is a multiple of 5
Returns 'FizzBuzz' if the input is a multiple of 3 and 5
Returns the input if it is not a multiple of 3 or 5
}
