% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzbuzz.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{Fizzbuzz}
\usage{
fizzbuzz(x)
}
\arguments{
\item{x}{\emph{Numeric} vector to run the fizzbuzz algorithm on}
}
\value{
A character vector containing numbers, \code{fizz}, \code{buzz} or \code{fizzbuzz}
}
\description{
Implements fizzbuzz
}
\examples{
fizzbuzz(1:10)
fizzbuzz(1)


}
