% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzbuzz.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{Fizzbuzz function}
\usage{
fizzbuzz(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Function implementing the fizzbuzz function.

The function takes a vector of numbers as inputs and replaces values that are
evenly divisible by 3 with "Fizz", values that are evenly divisible by 5 with "Buzz",
and values evenly divisible by both with "FizzBuzz".
}
\details{
The input vector is expected to be numeric and contain only finite values,
that are greater than or equal to 0 and must be representable as an integer. Any violations
of these assumptions will result in an error.
}
\examples{

fizzbuzz(1:3)
fizzbuzz(10:1)

\dontrun{
# The following will throw an error
fizzbuzz(-1)
}
}
