% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fizzbuzz.R
\name{fizzbuzz}
\alias{fizzbuzz}
\title{The fizzbuzz function}
\usage{
fizzbuzz(x)
}
\arguments{
\item{x}{A numeric vector of \emph{positive} finite values}
}
\value{
A character vector
}
\description{
Implements the fizzbuzz function for \strong{numeric} variables
}
\examples{
fizzbuzz(1:5)
fizzbuzz(5:1)
fizzbuzz(15)


}
